Alterações no script Visor
Alterações:
- Opções (Object, Mesh, Collider), podendo ser ativados e desativados individualmente (alterações feitas no inspetor).
- Quando o visor está ativo todos os objetos com script Visor sofrem suas respectivas ações.
- Quando o visor é desativado as ações são desfeitas.

+++ Assets/Game/Code/Scripts/Visor.cs
@@ -1,35 +1,86 @@
 using System.Collections.Generic;
-using System.Linq;
+using System.Runtime.CompilerServices;
 using UnityEngine;
 
 public class Visor : MonoBehaviour
 {
     [SerializeField] List<VisorTypes> Action;
 
+    private bool isVisorEnabled;
+    
+    private MeshRenderer meshRenderer;
+    private Collider collider;
+
+    private bool meshRendererEnabled;
+    private bool colliderEnabled;
+
     private void Start()
     {
-        Action.RemoveRange(3 , Action.Count - 3);
-        print(Action.Select(x => x.type));
+        if (Action.Count > 3) Action.RemoveRange(3 , Action.Count - 3);
+
+        isVisorEnabled = GameObject.Find("Visor");
+
+        meshRenderer = GetComponent<MeshRenderer>();
+        collider = GetComponent<Collider>();
+
+        meshRendererEnabled = meshRenderer != null ? GetComponent<MeshRenderer>().enabled : false;
+        colliderEnabled = collider != null ? GetComponent<Collider>().enabled : false;
+
+        DoAction();
     }
     void Update()
     {
-        /*if (GameObject.Find("Visor").activeSelf)
+        // Só realiza a ação a cada vez que o visor for ativado ou desativado. Melhor pra otimização.
+        if (GameObject.Find("Visor") != isVisorEnabled)
         {
-            switch (mode)
+            DoAction();
+            isVisorEnabled = GameObject.Find("Visor");
+        }
+    }
+
+    private void DoAction()
+    {
+        if (GameObject.Find("Visor") != null)
+        {
+            foreach (VisorTypes elem in Action)
             {
-                case Mode.Active:
-                    break;
+                switch (elem.type)
+                {
+                    case VisorTypes.Type.Object:
+
+                        if (meshRenderer) meshRenderer.enabled = elem.mode == VisorTypes.Mode.Active;
+                        if (collider) collider.enabled = elem.mode == VisorTypes.Mode.Active;
+
+                        break;
+
+                    case VisorTypes.Type.Mesh:
+
+                        if (meshRenderer) meshRenderer.enabled = elem.mode == VisorTypes.Mode.Active;
+
+                        break;
+                    case VisorTypes.Type.Collider:
+
+                        if (collider) collider.enabled = elem.mode == VisorTypes.Mode.Active;
+
+                        break;
+                }
             }
-        }*/
+        }
+        else
+        {
+            if (meshRenderer) meshRenderer.enabled = meshRendererEnabled;
+            if (collider) collider.enabled = colliderEnabled;
+        }
     }
 }
+
 
 [System.Serializable]
 public class VisorTypes
 {
-    public enum Type { Object, Mesh, Collider, sla, eae };
-    [Tooltip("What component will activate or desactivate when the visor is enable")] public Type type;
+    public enum Type { Object, Mesh, Collider };
+    [Tooltip("Which component will be enabled or disabled when the viewport is enabled. The Object option means (Mesh and Collider)")] public Type type;
 
-    public  enum Mode { Active, Inactive };
-    [Tooltip("The component will activate or desactivate when the visor is enable")] public Mode mode;
+    public enum Mode { Active, Inactive };
+    [Tooltip("The component will enabled or disabled when the visor is enable")] public Mode mode;
 }
